requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.9.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
